rimosso swipe left

+++ app/src/main/java/it/runningexamples/fiscalcode/SwipeCallback.java
@@ -4,7 +4,9 @@ import android.graphics.Canvas;
 import android.graphics.Color;
 import android.graphics.drawable.ColorDrawable;
 import android.graphics.drawable.Drawable;
+import android.util.Log;
 import android.view.View;
+import android.widget.Switch;
 
 import androidx.annotation.NonNull;
 import androidx.core.content.ContextCompat;
@@ -13,24 +15,20 @@ import androidx.recyclerview.widget.RecyclerView;
 
 public class SwipeCallback extends ItemTouchHelper.SimpleCallback {
 
-    private ColorDrawable redBackground, yellowBackground;
-    private Drawable deleteIcon, starIcon;
+    private ColorDrawable redBackground;
+    private Drawable deleteIcon;
     private RecyclerAdapter recyclerAdapter;
     private RecyclerView rcv;
-    int intrinsicWidthDelete, intrinsicHeightDelete, intrinsicWidthStar, intrinsicHeightStar;
+    int intrinsicWidthDelete, intrinsicHeightDelete;
 
     SwipeCallback(RecyclerAdapter adapter, RecyclerView rcv) {
-        super(ItemTouchHelper.ACTION_STATE_IDLE, ItemTouchHelper.LEFT|ItemTouchHelper.RIGHT);
+        super(ItemTouchHelper.ACTION_STATE_IDLE, ItemTouchHelper.LEFT);
         recyclerAdapter = adapter;
         this.rcv = rcv;
         redBackground = new ColorDrawable(Color.parseColor("#B80000"));
-        yellowBackground = new ColorDrawable(Color.YELLOW);
         deleteIcon = ContextCompat.getDrawable(recyclerAdapter.getContext(), R.drawable.bin72);
-        starIcon = ContextCompat.getDrawable(recyclerAdapter.getContext(), R.drawable.cityicon);
         intrinsicWidthDelete = deleteIcon.getIntrinsicWidth();
         intrinsicHeightDelete = deleteIcon.getIntrinsicHeight();
-        intrinsicWidthStar= starIcon.getIntrinsicWidth();
-        intrinsicHeightStar= starIcon.getIntrinsicHeight();
     }
 
     @Override
@@ -41,7 +39,13 @@ public class SwipeCallback extends ItemTouchHelper.SimpleCallback {
     @Override
     public void onSwiped(@NonNull RecyclerView.ViewHolder viewHolder, int direction) {
         int position = viewHolder.getAdapterPosition();
-        recyclerAdapter.deleteItem(position, rcv);
+        if (direction == ItemTouchHelper.LEFT){
+            Log.d("CodiceFiscale", String.valueOf(direction));
+            recyclerAdapter.deleteItem(position, rcv);
+        }else{
+            recyclerAdapter.notifyItemChanged(viewHolder.getAdapterPosition());
+            Log.d("CodiceFiscale", String.valueOf(direction));
+        }
     }
 
     @Override
@@ -57,23 +61,13 @@ public class SwipeCallback extends ItemTouchHelper.SimpleCallback {
         int deleteIconRight = itemView.getRight() - deleteIconMargin;
         int deleteIconBottom = deleteIconTop + intrinsicHeightDelete;
 
-        int starIconTop = itemView.getTop() + (itemHeight - intrinsicHeightStar) / 2;
-        int starIconMargin = (itemHeight - intrinsicHeightStar) / 8;
-        int starIconLeft = itemView.getRight() - starIconMargin -intrinsicHeightStar;
-        int starIconRight = itemView.getRight() - starIconMargin;
-        int starIconBottom = deleteIconTop + intrinsicHeightStar;
-
-        if (dX > 0){ //swipe verso destra
-            starIcon.setBounds(starIconLeft, starIconTop, starIconRight, starIconBottom);
-            yellowBackground.setBounds(itemView.getRight() + ((int)dX) - backgroundCornerOffset,
-            itemView.getTop(), itemView.getRight(), itemView.getLeft());
-        }
         if (dX < 0) { // swipe verso sinistra
             deleteIcon.setBounds(deleteIconLeft, deleteIconTop, deleteIconRight, deleteIconBottom);
             redBackground.setBounds(itemView.getRight() + ((int) dX) - backgroundCornerOffset,
                     itemView.getTop(), itemView.getRight(), itemView.getBottom());
         } else { // nessuno swipe
             redBackground.setBounds(0, 0, 0, 0);
+//            yellowBackground.setBounds(0,0,0,0);
         }
         redBackground.draw(c);
         deleteIcon.draw(c);
